# storage: optimize (*MemCachedStore).Persist for memory-backed ps